refactor: 카드 및 이미지 삭제 버튼에 대한 디렉토리 refactor
