+++ apps/admin/src/components/Button/Delete/Image/DeleteImageButton.styled.ts
