+++ apps/admin/src/components/Button/Delete/Image/DeleteImageButton.tsx
